refactor: remove indie dev
-Remove all code related to indie dev and confetti.

-Restructure routes for main game and words list to be more organized.
