+++ app/client/components/ui/interactive/GameDifficultyMenu.tsx
@@ -44,7 +44,7 @@ function GameDifficultyMenu({ title, emoji }: PropType) {
       </div>
       <div className="grid grid-cols-2 sm:grid-cols-4 lg:grid-cols-6 xl:grid-cols-8 gap-12 lg:gap-16 justify-center items-center">
         <button
-          onClick={() => navigate(`/word-skull-game-easy-mode`)}
+          onClick={() => navigate(`/game/word-skull-game-easy-mode`)}
           className="group relative flex h-full col-span-2 flex-col gap-2 w-[15em] text-pumpkin-orange hover:text-amber-600 justify-center items-center text-center pt-8 pb-12 hover:scale-[1.05] transition-scale duration-300 shadow-md hover:shadow-lg hover:shadow-amber-600/30 rounded-t-[150em] rounded-b-full bg-white shadow-pumpkin-orange/20 bg-opacity-10"
         >
           <div className="absolute bottom-2 shadow-inner rounded-full p-[0.08em] shadow-pumpkin-orange/20 group-hover:shadow-amber-600/30  opacity-65">{emoji}</div>
@@ -70,7 +70,7 @@ function GameDifficultyMenu({ title, emoji }: PropType) {
           </div>
         </button>
         <button
-          onClick={() => navigate(`/word-skull-game-medium-mode`)}
+          onClick={() => navigate(`/game/word-skull-game-medium-mode`)}
           className="group relative flex h-full col-span-2 flex-col gap-2 w-[15em] text-pumpkin-orange hover:text-amber-600 justify-center items-center text-center pt-8 pb-12  hover:scale-[1.05] transition-scale duration-300 shadow-md hover:shadow-lg hover:shadow-amber-600/30 rounded-t-[150em] rounded-b-full bg-white shadow-pumpkin-orange/20 bg-opacity-10"
         >
           <div className="absolute bottom-2 shadow-inner rounded-full p-[0.08em] shadow-pumpkin-orange/20 group-hover:shadow-amber-600/30  opacity-65">{emoji}</div>
@@ -96,7 +96,7 @@ function GameDifficultyMenu({ title, emoji }: PropType) {
           </div>
         </button>
         <button
-          onClick={() => navigate(`/word-skull-game-hard-mode`)}
+          onClick={() => navigate(`/game/word-skull-game-hard-mode`)}
           className="group relative flex h-full col-span-2 flex-col gap-2 w-[15em] text-pumpkin-orange hover:text-amber-600 justify-center items-center text-center pt-8 pb-12 hover:scale-[1.05] transition-scale duration-300 shadow-md hover:shadow-lg hover:shadow-amber-600/30 rounded-t-[150em] rounded-b-full bg-white shadow-pumpkin-orange/20 bg-opacity-10"
         >
           <div className="absolute bottom-2 shadow-inner rounded-full p-[0.08em] shadow-pumpkin-orange/20 group-hover:shadow-amber-600/30  opacity-65">{emoji}</div>
@@ -122,7 +122,7 @@ function GameDifficultyMenu({ title, emoji }: PropType) {
           </div>
         </button>
         <button
-          onClick={() => navigate(`/word-skull-game-extreme-mode`)}
+          onClick={() => navigate(`/game/word-skull-game-extreme-mode`)}
           className="group relative flex h-full col-span-2 flex-col lg:col-start-3 xl:col-start-auto gap-2 w-[15em] text-pumpkin-orange hover:text-amber-600 justify-center items-center text-center pt-8 pb-12  hover:scale-[1.05] transition-scale duration-300 shadow-md hover:shadow-lg hover:shadow-amber-600/30 rounded-t-[150em] rounded-b-full bg-white shadow-pumpkin-orange/20 bg-opacity-10"
         >
           <div className="absolute bottom-2 shadow-inner rounded-full p-[0.08em] shadow-pumpkin-orange/20 group-hover:shadow-amber-600/30  opacity-65">{emoji}</div>
