+++ app/routes/_index.tsx
@@ -51,7 +51,7 @@ function Header() {
         weapon at your disposal—your mind! The dungeon awaits...
       </p>
       <Link
-        to={"/word-skull-game-easy-mode"}
+        to={"/game/word-skull-game-easy-mode"}
         className={`hover:bg-amber-600 bg-pumpkin-orange flex z-10 text-white px-8  text-lg font-nunito rounded-full mt-3 py-2 tracking-widest leading-relaxed border-stone-300 hover:border-stone-400`}
       >
         Start Playing!
@@ -64,8 +64,6 @@ function Header() {
 }
 
 export default function Index() {
-  
-
   const { darkThemeActive } = useTheme();
 
   return (
@@ -101,7 +99,7 @@ export default function Index() {
               <p className="font-lato text-lg pl-5 tracking-wider leading-loose">
                 <Link
                   className={`text-pumpkin-orange hover:text-amber-600 font-lora`}
-                  to="/word-skull-game-easy-mode"
+                  to="/game/word-skull-game-easy-mode"
                 >
                   WordSkull
                 </Link>{" "}
