--- app/routes/best-indie-games-showcase.$game.tsx
@@ -1,7 +1,0 @@
-export default function BestIndieGameShowcase() {
-  return (
-    <div className="flex animate-fadeIn flex-col justify-center items-center mt-[3em]">
-      Here is the best indie game of the year
-    </div>
-  );
-}
