--- app/routes/edit-indie-game.tsx
@@ -1,239 +1,0 @@
-import {
-  ActionFunctionArgs,
-  Link,
-  LoaderFunctionArgs,
-  Outlet,
-  redirect,
-  useActionData,
-  useLocation,
-} from "react-router-dom";
-import { useTheme } from "../client/components/context/ThemeContext";
-import SocialLinks from "../client/components/navigation/SocialLinks";
-import IndieLoginForm from "../client/components/form/IndieLoginForm";
-import VerifyJWT from "../client/components/utils/validation/VerifyJWTIndieLogin";
-import PostIndieLogin from "../client/components/utils/requests/PostIndieLogin";
-import ProcessTryCatchErrors from "../client/components/utils/errors/ProcessErrors";
-
-interface ActionResponse {
-  error?: string;
-  message?: string;
-}
-
-export async function loader({ request }: LoaderFunctionArgs) {
-  try {
-    const currentUrl = new URL(request.url);
-    const jwtVerifiedUsername = await VerifyJWT({ request }); // Verify JWT token and retrieve username
-
-    // If JWT token exists, ensure the user has access to the submission form or redirect to 403
-    if (jwtVerifiedUsername) {
-      // If the user is trying to access the edit page but it's not their page, redirect to 403
-      if (
-        currentUrl.pathname !== "/edit-indie-game" &&
-        currentUrl.pathname !== `/edit-indie-game/${jwtVerifiedUsername}`
-      ) {
-        return redirect("/403");
-      }
-
-      // Redirect to user's specific edit page if they're visiting the general edit page
-      if (currentUrl.pathname === "/edit-indie-game") {
-        return redirect(`/edit-indie-game/${jwtVerifiedUsername}`);
-      }
-    } else {
-      // If no JWT token exists, redirect to the login page unless already on the login page
-      if (currentUrl.pathname !== "/edit-indie-game") {
-        return redirect("/edit-indie-game");
-      }
-    }
-
-    return {}; // If no action was taken, return an empty response
-  } catch (error) {
-    return ProcessTryCatchErrors({
-      error,
-      customError: "JWT verification failed",
-      status: 401,
-    });
-  }
-}
-
-export async function action({ request }: ActionFunctionArgs) {
-  const formData = await request.formData();
-  const username = formData.get("username")?.toString();
-  const password = formData.get("password")?.toString();
-
-  // Validate form fields
-  if (!username || !password) {
-    return new Response(JSON.stringify({ error: "All fields are required." }), {
-      status: 400,
-      headers: {
-        "Content-Type": "application/json",
-      },
-    });
-  }
-
-  return await PostIndieLogin({ username, password }); // Post IndieDev login credentials to API
-}
-
-export default function EditIndieGame() {
-  const actionData = useActionData() as ActionResponse;
-  const location = useLocation();
-  const { darkThemeActive } = useTheme();
-
-  return (
-    <div
-      className={`${
-        darkThemeActive ? "text-slate-300" : "text-skull-dark-brown"
-      } flex animate-fadeIn justify-center flex-col items-center mt-[3em]`}
-    >
-      <header className="flex flex-col justify-center items-center gap-3 mb-3 mx-5 text-center">
-        <h1
-          className={`${
-            darkThemeActive ? "text-slate-400" : "text-skull-dark-brown"
-          } w-full z-1  flex justify-center items-center flex-col md:flex-row text-5xl text-center mt-1 leading-snug -translate-y-[0.3em] sm:translate-y-0 font-lora tracking-wide`}
-        >
-          {location.pathname === "/edit-indie-game"
-            ? "Edit Your Game's Page"
-            : "Your Game's Page Header"}
-        </h1>
-      </header>
-      <main className="flex flex-col gap-5 justify-center items-center w-full max-w-[1200px]">
-        {location.pathname === "/edit-indie-game" ? (
-          <IndieLoginForm actionData={actionData} />
-        ) : (
-          <Outlet />
-        )}
-      </main>
-      <section className="flex relative flex-col gap-4 mt-14 mb-10 font-lato tracking-wider max-w-[1200px] text-lg leading-loose">
-        <div
-          className="absolute -top-[5em]"
-          id="indie-game-terms-of-service"
-         />
-        <h2 className="flex py-2 text-4xl font-lora text-center w-full justify-center items-center mb-2">
-          Terms and Conditions for Game Submission
-        </h2>
-        <p>
-          By submitting your indie game to be featured on{" "}
-          <Link
-            to="/"
-            className={`${
-              darkThemeActive ? "text-amber-600" : "text-pumpkin-orange"
-            } hover:text-amber-600 font-lora`}
-          >
-            WordSkull
-          </Link>
-          , you agree to the following terms and conditions:
-        </p>
-        <ul className="font-lato pl-7 flex gap-5 flex-col list-disc">
-          <li>
-            You will post original content that does not infringe any
-            third-party intellectual property rights.
-          </li>
-          <li>You will not post duplicate content from other websites.</li>
-          <li>
-            You will not post content that is illegal, threatening, abusive,
-            harassing, defamatory, vulgar, obscene, pornographic, or hateful.
-          </li>
-          <li>
-            You will not post content that is spammy, misleading, or deceptive.
-          </li>
-          <li>
-            You will not post content that violates any applicable laws or
-            regulations.
-          </li>
-          <li>You will not link to any obscene or illegal websites.</li>
-          <li>You will post content that is related to your game.</li>
-          <li>
-            If your content violates any of our terms or fails to meet our
-            standards, we reserve the right to modify or remove your content,
-            and account, without notice.
-          </li>
-        </ul>
-        <h2 className="flex py-2 text-4xl font-lora text-center w-full justify-center items-center mb-2">
-          Final Notes
-        </h2>
-        <p>
-          The header section is mandatory. If you haven&apos;t filled out the
-          header section of the form and submitted it, your game will not show
-          up on the{" "}
-          <Link
-            to="/best-indie-games-showcase"
-            className={`${
-              darkThemeActive ? "text-amber-600" : "text-pumpkin-orange"
-            } hover:text-amber-600 font-lora`}
-          >
-            Indie Games page
-          </Link>
-          , even if you save changes to other sections. Submissions may take
-          some time to process for certain sections so don&apos;t leave the page
-          until the submission is complete. All images are resized and optimized
-          for web before being uploaded to the back-end. They will likely be
-          resized to 640px by 360px, so a 16:9 aspect ratio is recommended for
-          best results. Ideally, your image should be in PNG or JPEG format. You
-          can experiment with other formats if you prefer, but there&apos;s a
-          chance it may be rejected. I&apos;ve added logic to handle GIFs which
-          should work in most cases, but I have not tested it extensively.
-        </p>
-        <p>
-          All games will be featured in random order on the{" "}
-          <Link
-            to="/best-indie-games-showcase"
-            className={`${
-              darkThemeActive ? "text-amber-600" : "text-pumpkin-orange"
-            } hover:text-amber-600 font-lora`}
-          >
-            Indie Games page
-          </Link>
-          , as well as, the sidebar on relevant pages. This way all contributors
-          get a fair chance at being featured. There will be no favoritism
-          towards any game on the{" "}
-          <Link
-            to="/best-indie-games-showcase"
-            className={`${
-              darkThemeActive ? "text-amber-600" : "text-pumpkin-orange"
-            } hover:text-amber-600 font-lora`}
-          >
-            Indie Games page
-          </Link>
-          .
-        </p>
-        <p>
-          Thank you for your submission. If the submission form is missing any
-          features, or if you have any questions or concerns, please don&apos;t
-          hesitate to{" "}
-          <Link
-            to="/socials"
-            className={`${
-              darkThemeActive ? "text-amber-600" : "text-pumpkin-orange"
-            } hover:text-amber-600 font-lora`}
-          >
-            contact us
-          </Link>
-          . I&apos;m excited to have your game showcased on{" "}
-          <Link
-            to="/"
-            className={`${
-              darkThemeActive ? "text-amber-600" : "text-pumpkin-orange"
-            } hover:text-amber-600 font-lora`}
-          >
-            WordSkull
-          </Link>
-          !
-        </p>
-        <p>
-          Once published, you can find your game on the{" "}
-          <Link
-            to="/best-indie-games-showcase"
-            className={`${
-              darkThemeActive ? "text-amber-600" : "text-pumpkin-orange"
-            } hover:text-amber-600 font-lora`}
-          >
-            indie game&apos;s featured page
-          </Link>
-          !
-        </p>
-      </section>
-      <section>
-        <SocialLinks />
-      </section>
-    </div>
-  );
-}
