+++ app/routes/game.tsx
@@ -1,0 +1,7 @@
+import { Outlet } from "react-router-dom";
+
+//Using this to group all game routes
+export default function Games() {
+  return <Outlet />;
+}
+
