+++ app/routes/z-words.tsx
@@ -1,0 +1,7 @@
+import { Outlet } from "react-router-dom";
+
+//Using this to group all words list items at the bottom of all routes
+export default  function ZWords() {
+  return <Outlet />;
+}
+
