+++ public/sitemap.xml
@@ -2,138 +2,138 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
     <loc>https://www.wordskull.com/</loc>
-    <lastmod>2025-01-05T20:25:56.130Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.522Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/about</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/all-words-for-word-game-eight-letters</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/all-words-for-word-game-five-letters</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/all-words-for-word-game-four-letters</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/all-words-for-word-game-nine-letters</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/all-words-for-word-game-seven-letters</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/all-words-for-word-game-six-letters</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/all-words-for-word-game-three-letters</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/all-words-for-word-game</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/best-indie-games-showcase.game</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/best-indie-games-showcase</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/blog.content</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/blog</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/board-game-reviews.reviews</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/board-game-reviews</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/cookies-policy</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/create-indie-showcase</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/edit-indie-game.username</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/edit-indie-game</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/faq</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/games</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/mdx/about-en</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/privacy-policy</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/project-milestones</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/sitemap</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/socials</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/terms-of-service</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/wallpaper</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/word-skull-game-easy-mode</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/word-skull-game-extreme-mode</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/word-skull-game-hard-mode</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/word-skull-game-medium-mode</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
   <url>
     <loc>https://www.wordskull.com/_index</loc>
-    <lastmod>2025-01-05T20:25:56.131Z</lastmod>
+    <lastmod>2025-01-05T22:14:24.523Z</lastmod>
   </url>
 </urlset>
